pyproject.toml
[project]
name = "taxitripproject"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "faker>=40.5.1,<41.0.0",
    "kafka-python>=2.3.0,<3.0.0",
    "pyspark (==3.5.8)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
